// tmp/test/models/CaptchaData.ts
/**
 * CNet API
 * CNet基础框架API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CaptchaData {
    'id'?: string | null;
    'backgroundImage'?: string | null;
    'sliderImage'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "backgroundImage",
            "baseName": "backgroundImage",
            "type": "string",
            "format": ""
        },
        {
            "name": "sliderImage",
            "baseName": "sliderImage",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CaptchaData.attributeTypeMap;
    }

    public constructor() {
    }
}