// tmp/test/models/LoginViewModel.ts
/**
 * CNet API
 * CNet基础框架API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class LoginViewModel {
    /**
    * 账号
    */
    'userName'?: string | null;
    /**
    * 密码
    */
    'password'?: string | null;
    /**
    * 密码键
    */
    'passwordKey'?: string | null;
    /**
    * 验证码Id
    */
    'captchaId'?: string | null;
    /**
    * 验证码数据
    */
    'captchaData'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "userName",
            "baseName": "userName",
            "type": "string",
            "format": ""
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string",
            "format": ""
        },
        {
            "name": "passwordKey",
            "baseName": "passwordKey",
            "type": "string",
            "format": ""
        },
        {
            "name": "captchaId",
            "baseName": "captchaId",
            "type": "string",
            "format": ""
        },
        {
            "name": "captchaData",
            "baseName": "captchaData",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return LoginViewModel.attributeTypeMap;
    }

    public constructor() {
    }
}