// tmp/test/models/Menu.ts
/**
 * CNet API
 * CNet基础框架API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Meta } from '../models/Meta';
import { HttpFile } from '../http/http';

export class Menu {
    /**
    * 路由路径
    */
    'path'?: string | null;
    /**
    * 名称
    */
    'name'?: string | null;
    'meta'?: Meta;
    /**
    * 子组件
    */
    'children'?: Array<Menu> | null;
    'component'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "path",
            "baseName": "path",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "meta",
            "baseName": "meta",
            "type": "Meta",
            "format": ""
        },
        {
            "name": "children",
            "baseName": "children",
            "type": "Array<Menu>",
            "format": ""
        },
        {
            "name": "component",
            "baseName": "component",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Menu.attributeTypeMap;
    }

    public constructor() {
    }
}