// tmp/test/models/Meta.ts
/**
 * CNet API
 * CNet基础框架API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Meta {
    /**
    * 图标
    */
    'icon'?: string | null;
    /**
    * 不缓存
    */
    'notCache'?: boolean | null;
    /**
    * 标题
    */
    'title'?: string | null;
    /**
    * 权限列表
    */
    'access'?: Array<string> | null;
    /**
    * 是否新窗口
    */
    'isNewWindow'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "icon",
            "baseName": "icon",
            "type": "string",
            "format": ""
        },
        {
            "name": "notCache",
            "baseName": "notCache",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "access",
            "baseName": "access",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "isNewWindow",
            "baseName": "isNewWindow",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Meta.attributeTypeMap;
    }

    public constructor() {
    }
}