// tmp/test/models/PubFunction.ts
/**
 * CNet API
 * CNet基础框架API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class PubFunction {
    'functionCode'?: string | null;
    'functionEnglish'?: string | null;
    'functionChina'?: string | null;
    'functionDescrip'?: string | null;
    'parentCode'?: string | null;
    'menuFlag'?: boolean;
    'stopFlag'?: boolean | null;
    'urlString'?: string | null;
    'editdate'?: Date | null;
    'editor'?: string | null;
    'sortidx'?: number | null;
    'target'?: string | null;
    'menuIcon'?: string | null;
    'routerPath'?: string | null;
    'isCache'?: boolean | null;
    'isNewWindow'?: boolean | null;
    'isOpen'?: boolean | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "functionCode",
            "baseName": "functionCode",
            "type": "string",
            "format": ""
        },
        {
            "name": "functionEnglish",
            "baseName": "functionEnglish",
            "type": "string",
            "format": ""
        },
        {
            "name": "functionChina",
            "baseName": "functionChina",
            "type": "string",
            "format": ""
        },
        {
            "name": "functionDescrip",
            "baseName": "functionDescrip",
            "type": "string",
            "format": ""
        },
        {
            "name": "parentCode",
            "baseName": "parentCode",
            "type": "string",
            "format": ""
        },
        {
            "name": "menuFlag",
            "baseName": "menuFlag",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "stopFlag",
            "baseName": "stopFlag",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "urlString",
            "baseName": "urlString",
            "type": "string",
            "format": ""
        },
        {
            "name": "editdate",
            "baseName": "editdate",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "editor",
            "baseName": "editor",
            "type": "string",
            "format": ""
        },
        {
            "name": "sortidx",
            "baseName": "sortidx",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "target",
            "baseName": "target",
            "type": "string",
            "format": ""
        },
        {
            "name": "menuIcon",
            "baseName": "menuIcon",
            "type": "string",
            "format": ""
        },
        {
            "name": "routerPath",
            "baseName": "routerPath",
            "type": "string",
            "format": ""
        },
        {
            "name": "isCache",
            "baseName": "isCache",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isNewWindow",
            "baseName": "isNewWindow",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isOpen",
            "baseName": "isOpen",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return PubFunction.attributeTypeMap;
    }

    public constructor() {
    }
}